Added teh OSNetworking package in libs directory.
Edited ipmodule.py, routemodule.py, and ifconfigmodule.py to
algin with new OSNetworking package.

ipmodule.py +35 : The ip bash command needs flags
ipmodule.py +38, +50 : Added wait to the flags
ipmodule.py +47 : The flags where wrong for ip this was corrected

routemodule.py +37-38 : Added wait as well as defaultKwargs.
This was done because it is unknown if the route command is installed on the box in question because it is a depercated command.
routemodule.py +43 : changed dataType

ifconfigmodule.py +36-48 : Added a new method to follow along with routemodule.py and ipmodule.py.