Adjusted the generic exception found in sshBufferControl.py's executeOnEnvironment method to raise the exception it found. This is actually how generic exceptions where handled originally. However, I added some more debug log information.
I removed the unnecessary while loop that waited for recv_ready. This was not supposed to be there at all as it invalidates the use of the '_bufferTimeToFirstBit' method and the 'firstBitTimeout'
variable. But in order to get excepted behavior out of 'unsafe' execution I added the use of the
'_bufferTimeToFirstBit' method to the unsafe code block.

I also removed some unnecessary comments.