Updated 'doesCommandExistPreParser' method to better control the command flow within a CommandContainer by calling setFailure if the command returns False.
This behavior is controlled by the 'PreParserStopOnFailure' currently only
passable directly to the 'doesCommandExistPreParser' method.

Added some additional logic to the '_preparserRunner' method within CC.
If there is a list of preparsers this will check the failure/complete class
variables as it iterates through the preparsers and will break if one or both
of the variables are true.

Updated unittesting to deal with depercated networking modules